Download QCEW dataset (size files 1st quarter) directly from the BLS website

Download QCEW dataset (size files 1st quarter) directly from the BLS website

download_qcew_size_data(target_year, industry = "naics", path_data = "./",
  url_wayback = "", unzip_dir = T)

Arguments

target_year
year for which we want to download the data

industry
download naics or sic data

path_data
where does the download happen: default current directory

url_wayback
allows to specify the path in internet wayback machine that kept some of the archive

unzip_dir
default is True, False is useful if you are just downloading the data

Value

read_list String with names of downloaded files

Note

Downloads the file to the current directory and unzips it.